% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DATASET}
\alias{DATASET}
\title{Two-stage example
Small example dataset for demonstrating matching, censoring imputation,
and treatment-rule learning functions in \pkg{DRMLSurv}.}
\format{
An object of class \code{Drmatch}
}
\usage{
data(DATASET)
}
\arguments{
\item{DATASET}{The data for moderation analysis}
}
\description{
Two-stage example
Small example dataset for demonstrating matching, censoring imputation,
and treatment-rule learning functions in \pkg{DRMLSurv}.
}
\examples{
head(data(DATASET))
}
\keyword{datasets}
