% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizer.R
\name{print.summary.Drmatch}
\alias{print.summary.Drmatch}
\title{Print a summary.Drmatch object}
\usage{
\method{print}{summary.Drmatch}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{summary.Drmatch()}.}

\item{...}{Additional arguments passed to \code{print()}.}
}
\description{
Print a summary.Drmatch object
}
