% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMSurv-package.R
\docType{package}
\name{RMSurv-package}
\alias{RMSurv}
\alias{RMSurv-package}
\title{RMSurv: Doubly Robust ML and Matching for Two-Stage Survival Outcomes}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Implements a two-stage survival analysis pipeline that combines (i) imputation of censored stage-specific survival times via donor matching, (ii) construction of matched counterfactual outcomes under static and dynamic treatment regimes, and (iii) estimation of optimal dynamic treatment rules using machine learning and cross-fitting. The toolkit supports nearest and full matching, optional exact matching, balance diagnostics, and flexible censoring/propensity score models via DoubleScore/SuperLearner-style learners.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ericanto0.github.io/RMSurv/}
  \item Report bugs at \url{https://github.com/EricAnto0/RMSurv/issues}
}

}
\author{
\strong{Maintainer}: Eric Anto \email{eric.anto@utah.edu} (\href{https://orcid.org/0000-0003-4971-949X}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Jincheng Shen \email{jincheng.shen@hsc.utah.edu} [copyright holder]
}

Other contributors:
\itemize{
  \item Tian Jiang \email{Tian.Jiang@utah.edu} [contributor]
}

}
\keyword{internal}
